requirements.txt
httpx~=0.28.1
aiohttp~=3.12.6
yarl~=1.20.0
colorlog~=6.9.0
requests~=2.32.3
pandas~=2.2.3
chardet~=5.2.0
tenacity~=9.1.2
netaddr~=1.3.0
eventlet~=0.40.1
tomli~=2.0.2
dotenv~=0.9.9
python-dotenv~=1.0.1
watchdog~=6.0.0
fastapi~=0.120.0
pydantic~=2.11.1
starlette~=0.48.0
aiocache~=0.12.3
uvicorn~=0.34.0
jinja2~=3.1.4
ulid-py~=1.1.0
redis~=5.2.0